Move GlueItemWidget to glue.utils.qt.mixins
