+++ glue/qt/data_collection_model.py
@@ -8,10 +8,10 @@ from glue.core.hub import HubListener
 from glue.core import message as m
 from glue.core.decorators import memoize
 from glue import core
-from glue.qt.mime import LAYERS_MIME_TYPE, PyMimeData
+from glue.qt.mime import LAYERS_MIME_TYPE
 from glue.qt.qtutil import layer_icon
 from glue.qt.widgets.style_dialog import StyleDialog
-
+from glue.utils.qt import PyMimeData
 
 DATA_IDX = 0
 SUBSET_IDX = 1
