+++ glue/qt/layer_artist_model.py
@@ -15,10 +15,10 @@ from glue.external.qt.QtCore import Qt
 from glue.external.qt import QtGui, QtCore
 from glue.clients.layer_artist import LayerArtistBase, LayerArtistContainer
 from glue.qt.qtutil import layer_artist_icon
-from glue.qt.mime import PyMimeData, LAYERS_MIME_TYPE
+from glue.qt.mime import LAYERS_MIME_TYPE
 from glue.qt.widgets.style_dialog import StyleDialog
 from glue.utils import nonpartial
-from glue.utils.qt import PythonListModel
+from glue.utils.qt import PythonListModel, PyMimeData
 
 
 class LayerArtistModel(PythonListModel):
